requirements.txt
# Core
aiopg==0.10.0
peewee==2.8.3
# peewee-async==0.5.5
git+https://github.com/05bit/peewee-async.git
git+https://github.com/CodeTeam/tcrudge
git+https://github.com/tarantool/tarantool-python.git@master
aiotarantool==1.0.6
psycopg2==2.6.2
tornado==4.4.2
jsonschema==2.5.1
msgpack-python==0.4.8
manage.py==0.2.10

# Migrations
django==1.10.2
django-environ==0.4.0

# Development & Testing
py==1.4.31
pytest==3.0.2
pytest-cov==2.3.1
pytest-env==0.6.0
pytest-tornado==0.4.5
coverage==4.2